Tweaks to get coap working with latest libraries
