+++ kafka-connect-coap/src/test/scala/com/datamountaineer/streamreactor/connect/coap/connection/TestDTLSConnection.scala
@@ -3,25 +3,18 @@ package com.datamountaineer.streamreactor.connect.coap.connection
 
 
 import com.datamountaineer.streamreactor.connect.coap.TestBase
-import com.datamountaineer.streamreactor.connect.coap.configs.{CoapSettings, CoapSinkConfig}
+import com.datamountaineer.streamreactor.connect.coap.configs.{CoapConstants, CoapSettings, CoapSinkConfig}
 import org.eclipse.californium.core.CoapClient
 import org.scalatest.BeforeAndAfter
+
+import java.util
+import scala.jdk.CollectionConverters.MapHasAsJava
 
 /**
   * Created by andrew@datamountaineer.com on 25/08/2017. 
   * stream-reactor
   */
 class TestDTLSConnection extends TestBase with BeforeAndAfter {
-
-//  val server = new Server(SINK_PORT_SECURE, SINK_PORT_INSECURE, kEY_PORT_SECURE)
-//
-//  before { server.start() }
-//  after { server.stop() }
-//
-//  CaliforniumLogger.initialize()
-//  CaliforniumLogger.setLevel(Level.INFO)
-//  ScandiumLogger.initialize()
-//  ScandiumLogger.setLevel(Level.INFO)
 
   "should create a DTSConnection with Public/Private PEM keys" in {
     val props = getPropsSecurePEM
@@ -38,4 +31,27 @@ class TestDTLSConnection extends TestBase with BeforeAndAfter {
     val client = DTLSConnectionFn(settings.head)
     client.isInstanceOf[CoapClient] shouldBe true
   }
+
+  def getPropsSecurePSK: util.Map[String, String] = {
+    Map(
+      CoapConstants.COAP_IDENTITY->"andrew",
+      CoapConstants.COAP_SECRET->"kebab",
+      CoapConstants.COAP_KCQL->SOURCE_KCQL_SECURE,
+      CoapConstants.COAP_URI->KEY_URI,
+      CoapConstants.COAP_DTLS_BIND_PORT->s"$kEY_PORT_SECURE"
+    ).asJava
+  }
+
+  def getPropsSecurePEM: util.Map[String, String] = {
+    Map(
+      CoapConstants.COAP_IDENTITY->"andrew",
+      CoapConstants.COAP_SECRET->"kebab",
+      CoapConstants.COAP_PRIVATE_KEY_FILE->PRIVATE_KEY_PATH,
+      CoapConstants.COAP_PUBLIC_KEY_FILE->PUBLIC_KEY_PATH,
+      CoapConstants.COAP_KCQL->SOURCE_KCQL_SECURE,
+      CoapConstants.COAP_URI->KEY_URI,
+      CoapConstants.COAP_DTLS_BIND_PORT->s"$kEY_PORT_SECURE"
+    ).asJava
+  }
+
 }
